feat: use novel Termificator by default in DSL
